application/: rename `app/` to be `src/`
