+++ application/alembic/env.py
@@ -3,8 +3,8 @@ from logging.config import fileConfig
 from alembic import context
 from sqlalchemy import engine_from_config, pool
 
-from app.config import config as app_config
-from app.models import BaseDBModel
+from src.config import config as app_config
+from src.models import BaseDBModel
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
