+++ application/pyproject.toml
@@ -38,7 +38,7 @@ line-length = 88
 line_length = 88
 multi_line_output = 3
 include_trailing_comma = true
-known_first_party = ["app", "tests"]
+known_first_party = ["src", "tests"]
 default_section = "THIRDPARTY"
 sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
 extend_skip = ["alembic"]
