+++ application/src/core/db.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm.session import Session
 
-from app.config import config
+from src.config import config
 
 engine = create_engine(config.get_postgres_uri(), pool_pre_ping=True)
