+++ application/src/core/exceptions.py
