+++ application/src/endpoints/dependencies/db.py
@@ -3,7 +3,7 @@ from typing import Annotated, Generator
 from fastapi import Depends
 from sqlalchemy.orm.session import Session
 
-from app.core.db import get_session
+from src.core.db import get_session
 
 
 def yield_session() -> Generator[Session, None, None]:
